test(stochtrace): Move shared nystrom fixture to conftest
