+++ tests/test_stochtrace/test_leave_one_out/conftest.py
@@ -1,6 +1,7 @@
 """Helpers for the leave-one-out experiments."""
 
-import numpy as np
+from matfree import stochtrace
+from matfree.backend import np, testing
 
 
 def exp_eigvals(n, /):
@@ -13,3 +14,12 @@ def step_eigvals(n, /, *, num_flat=50, drop_value=1e-3):
     eigvals_flat = np.ones(num_flat)
     eigvals_drop = np.ones(n - num_flat) * drop_value
     return np.concatenate([eigvals_flat, eigvals_drop])
+
+
+@testing.fixture(name="nystrom")
+@testing.parametrize(
+    "factory", [stochtrace.nystrom_eigh, stochtrace.nystrom_shifted_cholesky]
+)
+def fixture_nystrom(factory):
+    """Create a nystrom with default settings."""
+    return factory()
