+++ tests/test_stochtrace/test_leave_one_out/test_xnystrace.py
@@ -3,15 +3,6 @@
 from matfree import stochtrace, test_util
 from matfree.backend import func, linalg, np, prng, testing
 
-
-
-@testing.fixture(name="nystrom")
-@testing.parametrize(
-    "factory", [stochtrace.nystrom_eigh, stochtrace.nystrom_shifted_cholesky]
-)
-def fixture_nystrom(factory):
-    """Create a nystrom with default settings."""
-    return factory()
 from .conftest import exp_eigvals, step_eigvals
